migrating to neo4j 5, implementing TRAPI attribute constraints, removing compound qgraph handling
Updated cypher and all functionalities needed to migrate to neo4j 5, including support for bolt and http protocols.

This includes a significant change to handling results, instead of formatting results in the cypher statement do it in python. The main reasoning for this change was that the implicit aggregation used in the old return clauses are not supported by neo4j 5.

This also implements attribute constraints as they should be now, using the attributes in the proper TRAPI fields and not using arbitrary properties in nodes or edges as constraints. Only direct matches with == and === are supported as of now.