implements json/nested attributes, post-cypher transformation logic from Plater
It didn't make a lot of sense to have so much similar or redundant logic split across Plater and the transpiler. This moves most of the TRAPI formatting and attribute handling (including knowledge sources) and customization from Plater into the transpiler, consolidating and cleaning it up.

This also has the first implementation of the json string 'attributes' attribute, in order to support dictionaries or nested attributes as attributes in neo4j.