fixing potential issues with variable length edges
In rare cases where a specific edge is part of a variable length (subclass) edge and present in the result path,
an edge could be added twice if generated edge_ids are created for both instances. It would also be unpacked and transformed multiple times for no reason. This commit addresses both issues.